fix log kernel

+++ rwkv/triton/wkv/vanilla.py
@@ -73,16 +73,16 @@ def wkv_triton_vanilla_forward_kernel(
     beta_out_ptr = state_out_ptr + b_idx * state_out_s_b + state_out_s_ab
 
     # Loads parameters.
-    alpha = tl.load(alpha_ptr + cs * state_s_c, mask=cmask).to(tl.float32)
-    beta = tl.load(beta_ptr + cs * state_s_c, mask=cmask).to(tl.float32)
-    w = tl.load(w_ptr + cs * w_s_c, mask=cmask).to(tl.float32)
-    u = tl.load(u_ptr + cs * u_s_c, mask=cmask).to(tl.float32)
+    alpha = tl.load(alpha_ptr + cs * state_s_c, mask=cmask)
+    beta = tl.load(beta_ptr + cs * state_s_c, mask=cmask)
+    w = tl.load(w_ptr + cs * w_s_c, mask=cmask)
+    u = tl.load(u_ptr + cs * u_s_c, mask=cmask)
 
     ew = tl.exp(w)
 
     for t in range(tsz):
-        kt = tl.load(k_ptr + t * k_s_t + cs * k_s_c, mask=cmask).to(tl.float32)
-        vt = tl.load(v_ptr + t * v_s_t + cs * v_s_c, mask=cmask).to(tl.float32)
+        kt = tl.load(k_ptr + t * k_s_t + cs * k_s_c, mask=cmask)
+        vt = tl.load(v_ptr + t * v_s_t + cs * v_s_c, mask=cmask)
 
         euk = tl.exp(u + kt)
 
@@ -246,10 +246,10 @@ def wkv_vanilla_triton_backward_kernel(
     gbeta_out_ptr = gstate_out_ptr + b_idx * gstate_out_s_b + gstate_out_s_ab
 
     # Loads parameters.
-    galpha = tl.load(galpha_out_ptr + gstate_out_s_c * cs, mask=cmask).to(tl.float32)
-    gbeta = tl.load(gbeta_out_ptr + gstate_out_s_c * cs, mask=cmask).to(tl.float32)
-    w = tl.load(w_ptr + w_s_c * cs, mask=cmask).to(tl.float32)
-    u = tl.load(u_ptr + u_s_c * cs, mask=cmask).to(tl.float32)
+    galpha = tl.load(galpha_out_ptr + gstate_out_s_c * cs, mask=cmask)
+    gbeta = tl.load(gbeta_out_ptr + gstate_out_s_c * cs, mask=cmask)
+    w = tl.load(w_ptr + w_s_c * cs, mask=cmask)
+    u = tl.load(u_ptr + u_s_c * cs, mask=cmask)
 
     ew = tl.exp(w)
 
@@ -260,17 +260,19 @@ def wkv_vanilla_triton_backward_kernel(
     for t in range(tsz):
         tc = tsz - t - 1
 
-        kt = tl.load(k_ptr + tc * k_s_t + k_s_c * cs, mask=cmask).to(tl.float32)
-        vt = tl.load(v_ptr + tc * v_s_t + v_s_c * cs, mask=cmask).to(tl.float32)
-        alpha_prev = tl.load(alpha_ptr + tc * state_s_t + state_s_c * cs, mask=cmask).to(tl.float32)
-        beta_prev = tl.load(beta_ptr + tc * state_s_t + state_s_c * cs, mask=cmask).to(tl.float32)
+        kt = tl.load(k_ptr + tc * k_s_t + k_s_c * cs, mask=cmask)
+        vt = tl.load(v_ptr + tc * v_s_t + v_s_c * cs, mask=cmask)
+
+        alpha_prev = tl.load(alpha_ptr + tc * state_s_t + state_s_c * cs, mask=cmask)
+        beta_prev = tl.load(beta_ptr + tc * state_s_t + state_s_c * cs, mask=cmask)
+
         euk = tl.exp(u + kt)
         ek = tl.exp(kt)
 
         denom = beta_prev + euk
         denom_sq = denom * denom
 
-        gwkvt = tl.load(gwkv_ptr + tc * gwkv_s_t + gwkv_s_c * cs, mask=cmask).to(tl.float32)
+        gwkvt = tl.load(gwkv_ptr + tc * gwkv_s_t + gwkv_s_c * cs, mask=cmask)
 
         # Backpropagates wkv gradients.
         guk = gwkvt * euk * (beta_prev * vt - alpha_prev) / denom_sq
